add app file and loader component
